Sync QC_B with 64
git-svn-id: svn+ssh://192.168.1.9/var/svn/Frita/freeitathenscode@23 243fd332-3cb7-4356-b4cc-f2eb9687c9df

+++ QC_process_dev/Master_32/QC_Backend.sh
@@ -1,6 +1,6 @@
 #/bin/bash
 cd ||exit 3
-# *--* Treat references to undeclared variables as an error
+# *--* (set -u) = Treat references to undeclared variables as an error
 set -u
 # *--* Ensure dialog program is installed.
 type dialog &>/dev/null || {
@@ -18,123 +18,137 @@ for GL in $CPUFLAGS ;do if [ $GL == 'lm' ];then CPU_ADDRESS=64;fi;done
 # *--* Create log file
 rm QC.log* 2> /dev/null
 touch QC.log || exit 5
+
 # *--* Optical drive(s) QC_test
 TARGET_OPTICAL=''
-for QCVAR in ls -d /sys/block/sr*
+optical_drive_count=0
+for dev_path in $(ls -d /sys/block/sr*)
 do
-    if [ -z $TARGET_OPTICAL ]
-    then
-        echo "INFO : CD/DVD drive test. Too many optical drives exist!" >> QC.log
-        break
-    fi
-    [[ -z $TARGET_OPTICAL ]] && TARGET_OPTICAL=$QCVAR
+    dev_name=$(basename $dev_path)
+    [[ -z $TARGET_OPTICAL ]] && TARGET_OPTICAL=$dev_name
+    ((optical_drive_count++))
 done
-if [ -z $TARGET_OPTICAL ];then
+if   [ $optical_drive_count -lt 1 ]
+then
     echo "PROBLEM : CD/DVD drive test. Add an optical drive!" >> QC.log
-else 
+else
+    if [ $optical_drive_count -gt 1 ]
+    then
+        echo "INFO : CD/DVD drive test. More than one optical drive!" >> QC.log
+    fi
     echo "PASSED  : CD/DVD drive test (Have at least one drive)" >> QC.log
+fi
+if [ ! -z $TARGET_OPTICAL ];then
     TARGET_DEVICE="/dev/${TARGET_OPTICAL}"
-    eject $TARGET_DEVICE;RC=$?
+    SUBSH_SIG=$(mktemp -t "QC_B_subshell_msg.XXXXX" || /tmp/blah)
+    eject -T $TARGET_DEVICE;RC=$?
     if [ $RC -eq 0 ];then
-        #TODO special handling for laptops
-        (sleep 8 && eject -t $TARGET_DEVICE) &
-        dialog --title "Free IT Athens Quality Control Test"\
-            --pause "remove any Frita CDs (I'll try to close the drive after ~8 seconds...)" 8 90 8;clear
-        if [ $etRC -gt 0 ];then
-            dialog --title "Uh Oh!" --pause "Please close the optical drive manually.Thanks."\
-             12 50 5
-        fi;clear)
-
+        (sleep 8;eject -T $TARGET_DEVICE || echo '(Laptop??) Please close cd drive manually.'>$SUBSH_SIG) &
         dialog --title "Free IT Athens Quality Control Test"\
             --pause "remove any Frita CDs (I'll try to close the drive after ~8 seconds...)" 8 90 8;clear
     else
-        echo "PROBLEM: Cannot eject Optical Drive at $TARGET_DEVICE" >>QC.log
+        echo "PROBLEM: Cannot open Optical Drive at $TARGET_DEVICE" >>QC.log
     fi
+else
+    echo "PROBLEM: Internal error identifying optical drive." >>QC.log
 fi
-    #$sleepeject_PID=$!
-    #(sleep 8 && eject -t /dev/sr0) &
-
-# *--* hdd
-QCVAR=$(ls -d /sys/block/sd[a-z] |wc -l)
-if test $QCVAR -eq 1;then
-    echo "PASSED  : Hard drive test." >> QC.log
-elif test $QCVAR -gt 1;then
-    echo "WARNING : Hard drive test. Check that there is only one hard drive." >> QC.log
-elif test $QCVAR -lt 1;then
-    echo "PROBLEM : Hard drive test. Something went wrong with the test!" >> QC.log
-fi
+#$sleepeject_PID=$!
 
 # *--* network
-QCVAR=$(ls /sys/class/net | grep eth | wc -l)
-if test $QCVAR -lt 1
-then echo "PROBLEM : Network card test. There is no network card!" >> QC.log
-elif test $QCVAR -gt 1
-then echo "PROBLEM : Network card test. There are too many network cards!" >> QC.log
+dev_count=$(ls /sys/class/net | grep eth | wc -l)
+if   test $dev_count -lt 1;then 
+    echo "PROBLEM : Network card test. There is no network card!" >> QC.log
+elif test $dev_count -gt 1;then
+    echo "PROBLEM : Network card test. There are too many network cards!" >> QC.log
 else
-echo "PASSED  : Network card test." >> QC.log
+    echo "PASSED  : Network card test." >> QC.log
 fi
 
 # *--* modem detection
-QCVAR=$(lspci | grep -i Modem | wc -l)
-if test $QCVAR -ge 1
-then echo "PROBLEM : Modem test. Remove a modem from the computer!" >> QC.log
-else echo "PASSED  : Modem test." >> QC.log
+dev_count=$(lspci | grep -i Modem | wc -l)
+if test $dev_count -ge 1;then
+    echo "PROBLEM : Modem test. Remove a modem from the computer!" >> QC.log
+else
+    echo "PASSED  : Modem test." >> QC.log
 fi
 
 # *--* sound
-QCVAR=$(ls /sys/class/sound/ | grep card | wc -l)
-if test $QCVAR -lt 1
-then echo "PROBLEM : Sound card test. There is no sound card!" >> QC.log
-elif test $QCVAR -gt 1
-then echo "WARNING : Sound card test. Check that there is only one sound card." >> QC.log
+dev_count=$(ls /sys/class/sound/ | grep card | wc -l)
+if   test $dev_count -lt 1;then
+    echo "PROBLEM : Sound card test. There is no sound card!" >> QC.log
+elif test $dev_count -gt 1;then
+    echo "WARNING : Sound card test. Check that there is only one sound card." >> QC.log
 else
-echo "PASSED  : Sound card test." >> QC.log
+    echo "PASSED  : Sound card test." >> QC.log
 fi
 
 # *--* video
-QCVAR=$(ls /sys/class/graphics/ | grep fb[0-9] | wc -l)
-if test $QCVAR -lt 1
-then echo "PROBLEM : Video card test. Something went wrong with the test!" >> QC.log
-elif test $QCVAR -gt 1
-then echo "PROBLEM : Video card test. There are too many video cards in the computer!" >> QC.log
+dev_count=$(ls /sys/class/graphics/ | grep fb[0-9] | wc -l)
+if   test $dev_count -lt 1;then
+    echo "PROBLEM : Video card test. Something went wrong with the test!" >> QC.log
+elif test $dev_count -gt 1;then
+    echo "PROBLEM : Video card test. There are too many video cards in the computer!" >> QC.log
 else
-echo "PASSED  : Video card test." >> QC.log
+    echo "PASSED  : Video card test." >> QC.log
 fi
 
 # *--* resolution
 QCVAR=$(xrandr | grep '1024x768')
-if test -z "$QCVAR"
-then echo "PROBLEM : Video resolution test. Resolution must be at least 1024x768!" >> QC.log
+if test -z "$QCVAR";then
+    echo "PROBLEM : Video resolution test. Resolution must be at least 1024x768!" >> QC.log
 else
-echo "PASSED  : Video resolution test." >> QC.log
+    echo "PASSED  : Video resolution test." >> QC.log
 fi
 
 # *--* usb
-QCVAR=$(ls /sys/bus/usb/devices | wc -l)
-if test $QCVAR -lt 1
-then echo "PROBLEM : USB port test. There are no USB ports!" >> QC.log
+dev_count=$(ls /sys/bus/usb/devices | wc -l)
+if test $dev_count -lt 1;then
+    echo "PROBLEM : USB port test. There are no USB ports!" >> QC.log
 else
-echo "PASSED  : USB port test." >> QC.log
+    echo "PASSED  : USB port test." >> QC.log
 fi
 
 # *--* users
-QCVAR=$(ls /home | wc -l)
-if test $QCVAR -lt 1
-then echo "PROBLEM : User count. Something is wrong with this test!" >> QC.log
-elif test $QCVAR -gt 1
-then echo "PROBLEM : User count. There is more than one user account!" >> QC.log
+user_count=$(ls /home | wc -l)
+if   test $user_count -lt 1;then
+    echo "PROBLEM : User count. Something is wrong with this test!" >> QC.log
+elif test $user_count -gt 1;then
+    echo "PROBLEM : User count. There is more than one user account!" >> QC.log
 else
-echo "PASSED  : User count test." >> QC.log
+    echo "PASSED  : User count test." >> QC.log
 fi
 
 # *--* CPU speed
 QCVAR=$(awk '/MHz/ {print $4; exit}' /proc/cpuinfo)
 LEN=$(expr match $QCVAR '[0-9]*')
 QCVAR=${QCVAR:0:$LEN}
-if test $QCVAR -lt 650
-then echo "PROBLEM : CPU clockspeed test. Recycle this computer!" >> QC.log
+if test $QCVAR -lt 650;then
+    echo "PROBLEM : CPU clockspeed test. Recycle this computer!" >> QC.log
 else
     echo "PASSED  : CPU clockspeed test." >> QC.log
+fi
+
+# *--* Hard Drive(s) *--*
+prime_disk=''
+prime_sectors=0
+sdx_sectors=0
+sdx_count=0
+for sdx in $(ls -d /sys/block/sd[a-w])
+do
+    sdx_sectors=$(cat ${sdx}/size)
+    if [ $sdx_sectors -gt 0 ]
+    then
+        [[ -z "$prime_disk" ]] && prime_disk=$sdx
+        [[ $prime_sectors -eq 0 ]] && prime_sectors=$sdx_sectors
+        ((sdx_count++))
+    fi
+done
+if   test $sdx_count -eq 1;then
+    echo "PASSED  : Hard drive test." >> QC.log
+elif test $sdx_count -gt 1;then
+    echo "WARNING : Hard drive test. Check that there is only one hard drive." >> QC.log
+elif test $sdx_count -lt 1;then
+    echo "PROBLEM : Hard drive test. Something went wrong with the test!" >> QC.log
 fi
 
 PROCESSORS=$(grep 'physical id' /proc/cpuinfo | sort -u | wc -l)
@@ -142,8 +156,8 @@ CORES=$(grep 'core id' /proc/cpuinfo | sort -u | wc -l)
 
 if test $PROCESSORS -gt 1 -o $CORES -gt 1
 then
-    FS_LOW_VALUE=70000
-    #FS_LOW_VALUE=65200   #65271 for forgiving image part
+    # Treat as Dual-core
+    FS_LOW_VALUE=76000
     FS_HIGH_VALUE=1000000
     FS_TEXT="80GB to 1TB"
     # up to 128MB of shared memory for video
@@ -151,9 +165,10 @@ then
     RAM_HIGH_VALUE=$(expr 2048 \* 1024)
     RAM_TEXT="2GB"
 else
-    FS_LOW_VALUE=5001
+    # Single-core
+    FS_LOW_VALUE=38000
     FS_HIGH_VALUE=80000
-    FS_TEXT="10 to 80GB"
+    FS_TEXT="40 to 80GB"
     # up to 32MB of shared memory for video
     RAM_LOW_VALUE=$(expr 480 \* 1024)
     RAM_HIGH_VALUE=$(expr 1024 \* 1024)
@@ -161,10 +176,19 @@ else
 fi
 
 # *--* filesystem size
-QCVAR=$(df -m / | awk '/dev/ {print $4}')
-if test $QCVAR -lt "$FS_LOW_VALUE" -o $QCVAR -gt "$FS_HIGH_VALUE"
-then echo "PROBLEM : Free space test. Hard drive should be $FS_TEXT." >> QC.log
-else echo "PASSED  : Free space test." >> QC.log
+#   160041885696 is 160GB in bytes at least for some drives
+total_disk_bytes=$(echo "${prime_sectors}*$(cat $prime_disk/queue/hw_sector_size)" |bc)
+#TEST
+echo 'Disk Bytes' $total_disk_bytes >&2
+#ENDT
+QCVAR=$(echo "(($total_disk_bytes/1024)/1024)" |bc)
+#TEST
+echo 'Disk Megabytes' $QCVAR >&2
+#ENDT
+if test $QCVAR -lt "$FS_LOW_VALUE" -o $QCVAR -gt "$FS_HIGH_VALUE";then
+    echo "PROBLEM : Free space test. Hard drive should be $FS_TEXT." >> QC.log
+else
+    echo "PASSED  : Free space test." >> QC.log
 fi
 
 # *--* RAM
@@ -200,39 +224,73 @@ then
             fi
             ;;
         Disable_3D)
-            dialog 'Click Disable 3D Icon';;
-            dialog --title "Free IT Athens Quality Control Test" --textbox 'Click Disable 3D Icon' 20 80
+            dialog 'Click Disable 3D Icon'
+            dialog --title "Free IT Athens Quality Control Test" --msgbox "Click Disable 3D Icon" 20 80
+        ;;
     esac
 fi
 if [ ! -e $Lock_file ]
 then
-    echo "10 second long 3D test started" | tee $Lock_file
-    # run a 3D screensaver in a window for 10 seconds then stop it
-    /usr/lib/xscreensaver/antspotlight -window &
-    PID=$!
-    #dialog --title "Free IT Athens Quality Control Test"\
-    #    --pause '10 second 3D test started' 8 90 10;clear
-    sleep 10
-    kill $PID
-    # if the computer doesnt hang, it passes
-    rm -f $Lock_file
-    echo "PASSED  : 3D stability test." >> QC.log
+  if [ -f /usr/lib/xscreensaver/antspotlight ];then
+      echo "10 second long 3D test started" | tee $Lock_file
+      # run a 3D screensaver in a window for 10 seconds then stop it
+      /usr/lib/xscreensaver/antspotlight -window 2>/tmp/xscreensaver_test &
+      PID=$!
+      sleep 10
+      kill $PID
+      # if the computer doesnt hang, it passes
+      rm -f $Lock_file
+      echo "PASSED  : 3D stability test." >> QC.log
+  else
+      echo "WARNING: 3D stability test is not possible" >>QC.log
+  fi
 fi
 
+# *--* Test playing flash content
+path2firefox=$(which firefox 2>/dev/null)
+if [ ! -z "$path2firefox" ]
+then
+#dialog --title "Free IT Athens Quality Control Test" --msgbox "Shall I test shockwave flash in your browser - $path2firefox ?" 50 90
+Test_ff_msg=
+dialog --title "Free IT Athens Quality Control Test" --msgbox "Now testing shockwave flash in your browser - $path2firefox " 50 90
+$path2firefox -no-remote 'http://www.youtube.com/watch?v=7OXiS4BTXNQ' 2>/tmp/ff.err &
+ice_PID=$!;echo $ice_PID 'process # for ff'
+(sleep 20;kill $ice_PID) &
+fi
 # *--* sort to make problems more visible
 sort -r QC.log > QC.sorted.log
-if [ $CPU_ADDRESS == '32' ]
+if [ $CPU_ADDRESS -eq 32 ]
 then
-    #echo 'This is a 32 bit box (running XFCE)'
+    echo -e "This has a 32 bit CPU." |tee -a QC.sorted.log
     echo 'Remember to save the first XFCE session for the new user!' >>QC.sorted.log
+else
+    echo -e "CPU is 64-bit capable." |tee -a QC.sorted.log
+    if [ 0 -eq $(uname -mpi |grep x86_64 |wc -l) ]
+    then
+	echo "You MIGHT want to re-install using a 64-bit kernel." |tee -a QC.sorted.log
+    fi
 fi
+# Check for manual optical drive close message (e.g., laptops)
+[[ 0 -lt $(wc -l $SUBSH_SIG|cut -f1 -d' ') ]] && cat $SUBSH_SIG >>QC.sorted.log
 
 #output log to dialog box for ease of reading
 dialog --title "Free IT Athens Quality Control Test Results" --textbox QC.sorted.log 20 80
 clear
+
+# *--* QC_Backend.sh Finished *--*
+
 #TODO (for build) include tty fonts on libreoffice (or instructions)
 #TODO Need test for flash content handling
 #TODO Need change build to make separate partition for /home
-# Note to geekaholics: 
-#     existence of : /home/"Newuser"/.config/xfce4/xfconf/xfce-perchannel-xml/xfce4-session.xml:    <property name="SessionName" type="string" value="Default"/>
+#QCVAR=$(ls /sys/block/ | grep sr | wc -l)
+    #dialog --title "Free IT Athens Quality Control Test"\
+    #--pause "remove any Frita CDs (I'll try to close the drive after ~8 seconds...)" 8 90 8;clear
+#ls /sys/block/ | grep sr | wc -l)
+#if test $QCVAR -eq 1
+#then echo "PASSED  : CD/DVD drive test" >> QC.log
+#elif test $QCVAR -gt 1
+#then echo "PROBLEM : CD/DVD drive test. Too many optical drives exist!" >> QC.log
+#elif test $QCVAR -lt 1
+#then echo "PROBLEM : CD/DVD drive test. Add an optical drive!" >> QC.log
+#QCVAR=$(df -m / | awk '/dev/ {print $4}')
